Prank App.Visit Google Play for more .
Ever wanted to scare your friends with the sound of guns? Well, now you
can with Gun Sounds for Android! Watch as your friends jump, duck and
look for cover once they hear the frightening sound of realistic gunfire
- it’s a great prank to pull on everyone! ?
This app is perfect to use as a prank - from taking revenge on your
roommate or causing mass-terror when in a group!
? ? ? Unleash your inner-menace - with our Gun Sounds soundboard.